% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magic.R
\name{as.data.frame.magic}
\alias{as.data.frame.magic}
\title{Convert a MAGIC object to a data.frame}
\usage{
\method{as.data.frame}{magic}(x, ...)
}
\arguments{
\item{x}{A fitted MAGIC object}

\item{...}{Arguments for as.data.frame()}
}
\description{
Returns the smoothed data matrix
}
