% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magic.R
\name{as.matrix.magic}
\alias{as.matrix.magic}
\title{Convert a MAGIC object to a matrix}
\usage{
\method{as.matrix}{magic}(x, ...)
}
\arguments{
\item{x}{A fitted MAGIC object}

\item{...}{Arguments for as.matrix()}
}
\description{
Returns the smoothed data matrix
}
