% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_pymagic_version}
\alias{check_pymagic_version}
\title{Check that the current MAGIC version in Python is up to date.}
\usage{
check_pymagic_version()
}
\description{
Check that the current MAGIC version in Python is up to date.
}
