% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{install.magic}
\alias{install.magic}
\title{Install MAGIC Python Package}
\usage{
install.magic(envname = "r-reticulate", method = "auto",
  conda = "auto", pip = TRUE, ...)
}
\arguments{
\item{envname}{Name of environment to install packages into}

\item{method}{Installation method. By default, "auto" automatically finds
a method that will work in the local environment. Change the default to
force a specific installation method. Note that the "virtualenv" method
is not available on Windows.}

\item{conda}{Path to conda executable (or "auto" to find conda using the PATH
and other conventional install locations).}

\item{pip}{Install from pip, if possible.}

\item{...}{Additional arguments passed to conda_install() or
virtualenv_install().}
}
\description{
Install MAGIC Python package into a virtualenv or conda env.
}
\details{
On Linux and OS X the "virtualenv" method will be used by default
("conda" will be used if virtualenv isn't available). On Windows,
the "conda" method is always used.
}
