% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magic_testdata.R
\docType{data}
\name{magic_testdata}
\alias{magic_testdata}
\title{Fake scRNAseq data for examples}
\format{
A matrix with 500 rows and 197 variables
}
\source{
The authors
}
\usage{
magic_testdata
}
\description{
A subsampled dataset of epithelial to mesenchymal transition
}
\keyword{datasets}
