% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magic.R
\name{print.magic}
\alias{print.magic}
\title{Print a MAGIC object}
\usage{
\method{print}{magic}(x, ...)
}
\arguments{
\item{x}{A fitted MAGIC object}

\item{...}{Arguments for print()}
}
\description{
This avoids spamming the user's console with a list of many large matrices
}
\examples{
if (reticulate::py_module_available("magic")) {

data(magic_testdata)
data_magic <- magic(magic_testdata)
print(data_magic)
## MAGIC with elements
## $result : (500, 197)
## $operator : Python MAGIC operator
## $params : list with elements (data, knn, decay, t, npca, knn.dist.method)

}
}
