% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pymagic_is_available}
\alias{pymagic_is_available}
\title{Check whether MAGIC Python package is available and can be loaded}
\usage{
pymagic_is_available()
}
\description{
This is used primarily to avoid running tests on CRAN
and elsewhere where the Python package should not be
installed.
}
