% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magic.R
\name{summary.magic}
\alias{summary.magic}
\title{Summarize a MAGIC object}
\usage{
\method{summary}{magic}(object, ...)
}
\arguments{
\item{object}{A fitted MAGIC object}

\item{...}{Arguments for summary()}
}
\description{
Summarize a MAGIC object
}
\examples{
if (pymagic_is_available()) {
  data(magic_testdata)
  data_magic <- magic(magic_testdata)
  summary(data_magic)
  ## ZEB1
  ## Min.   :0.01071
  ## 1st Qu.:0.01119
  ## Median :0.01130
  ## Mean   :0.01129
  ## 3rd Qu.:0.01140
  ## Max.   :0.01201
}
}
